requirements.txt
pandas
sqlalchemy
pyodbc
python-dotenv

